LS_Backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
mongoengine==0.27.0
bcrypt==4.1.1
pymongo==4.6.0
